Move constraint designs into ConstraintExecution
